Don't reset total counts on generation change
To actually use total counts for metrics, those should not be
reset when the generation changes. This allows metrics services
to use gauges and time series database rate functionality for
visualizations and alerts. If the breaker is flipping between
open and half-open (or between any other states), resetting all
counters means metrics would most of the time say no requests
are coming in at all as all counters are zeroed.